신종하 : 플레이레벨 턴타임이 대기실 옵션셋팅에 따라 바뀌도록 수정

+++ GameLogic/GameOptionInfo.cpp
@@ -4,7 +4,7 @@
 
 // 턴타임
 // TernTime = 0 무한대(턴제한없음)
-int GameOptionInfo::TernTime = 45;
+float GameOptionInfo::TernTime = 45.f;
 
 // 라운드타임
 int GameOptionInfo::RoundTime = 15;
